Move scriptCmd to new Stack.Script module
